# lib/api-zod/src/generated/types/complaint.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CivicPulse: Smart Governance Platform for Administrative Operations with Citizen Service Assistance
 * OpenAPI spec version: 0.1.0
 */
import type { ComplaintStatus } from './complaintStatus';
import type { VerificationStatus } from './verificationStatus';

export interface Complaint {
  id: number;
  reference: string;
  title: string;
  category: string;
  description: string;
  location: string;
  /** @nullable */
  latitude: number | null;
  /** @nullable */
  longitude: number | null;
  /** @nullable */
  photoData?: string | null;
  status: ComplaintStatus;
  citizenId: number;
  citizenName: string;
  departmentId: number;
  departmentName: string;
  /** @nullable */
  assignedStaffId: number | null;
  /** @nullable */
  assignedStaffName: string | null;
  /** @nullable */
  remarks: string | null;
  /** @nullable */
  resolution: string | null;
  verificationStatus: VerificationStatus;
  /** @nullable */
  verifiedAt: string | null;
  /** @nullable */
  reopenReason: string | null;
  createdAt: string;
  updatedAt: string;
}