# lib/api-zod/src/generated/types/complaintInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CivicPulse: Smart Governance Platform for Administrative Operations with Citizen Service Assistance
 * OpenAPI spec version: 0.1.0
 */

export interface ComplaintInput {
  /** @minLength 3 */
  title: string;
  /** @minLength 2 */
  category: string;
  departmentId: number;
  /** @minLength 10 */
  description: string;
  /**
     * Deprecated compatibility field; the server derives the stored location from map coordinates.
     * @maxLength 180
     */
  location?: string;
  /**
     * @minimum -90
     * @maximum 90
     */
  latitude: number;
  /**
     * @minimum -180
     * @maximum 180
     */
  longitude: number;
  /**
     * @maxLength 7000000
     * @nullable
     */
  photoData?: string | null;
}