# lib/api-zod/src/generated/types/emergingIssue.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CivicPulse: Smart Governance Platform for Administrative Operations with Citizen Service Assistance
 * OpenAPI spec version: 0.1.0
 */

export interface EmergingIssue {
  category: string;
  area: string;
  recentComplaints: number;
  increasePercent: number;
  risk: string;
}