# lib/api-zod/src/generated/types/history.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CivicPulse: Smart Governance Platform for Administrative Operations with Citizen Service Assistance
 * OpenAPI spec version: 0.1.0
 */

export interface History {
  id: number;
  eventType: string;
  description: string;
  /** @nullable */
  performedBy: number | null;
  /** @nullable */
  performedByName: string | null;
  eventAt: string;
}