# lib/api-zod/src/generated/types/impactBreakdown.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CivicPulse: Smart Governance Platform for Administrative Operations with Citizen Service Assistance
 * OpenAPI spec version: 0.1.0
 */

export interface ImpactBreakdown {
  severity: number;
  affectedCitizens: number;
  recurrence: number;
  age: number;
  location: number;
}