# lib/api-zod/src/generated/types/impactScore.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CivicPulse: Smart Governance Platform for Administrative Operations with Citizen Service Assistance
 * OpenAPI spec version: 0.1.0
 */
import type { ImpactBreakdown } from './impactBreakdown';

export interface ImpactScore {
  impactScore: number;
  breakdown: ImpactBreakdown;
}