# lib/api-zod/src/generated/types/verificationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CivicPulse: Smart Governance Platform for Administrative Operations with Citizen Service Assistance
 * OpenAPI spec version: 0.1.0
 */

export interface VerificationInput {
  accepted: boolean;
  /**
     * @minLength 5
     * @maxLength 1000
     * @nullable
     */
  reason?: string | null;
}